Everything looks good.

+++ src/main/java/com/rosteringester/fileread/XLSXFile.java
@@ -114,7 +114,7 @@ public class XLSXFile extends Excel implements FileReader {
                                     xlsxData[rowTracker][colTracker] = df.format(cell.getDateCellValue());
                                     //System.out.println("[DATE]: " + df.format(cell.getDateCellValue()));
                                 } else {
-                                    xlsxData[rowTracker][colTracker] = String.valueOf((int) cell.getNumericCellValue()).toString();
+                                    xlsxData[rowTracker][colTracker] = String.valueOf((long) cell.getNumericCellValue()).toString();
                                     //System.out.println("[NUMERIC]: " + xlsxData[rowTracker][colTracker]);
                                 }
